Refactoring evaluator test I wanted to do this because I felt that showing some "hand.royal_flush" does not really show well what a royal flush is in the test. But I wanted some conviences for building out the card.
To make building the card more convienent, I updated the the factory where if a suit and literal value is given, a card like that is made. If it is anything else, the card factory is used and the values are merged.

Also if just a literal value is given, a card of any suit with that value should be given back.